Move clients/ to api/clients/
