+++ api/clients/cert_test.go
@@ -1,4 +1,4 @@
-package client
+package clients
 
 import (
 	"testing"
