+++ tools/traffic/generator_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Layr-Labs/eigenda/clients"
-	clientsmock "github.com/Layr-Labs/eigenda/clients/mock"
+	"github.com/Layr-Labs/eigenda/api/clients"
+	clientsmock "github.com/Layr-Labs/eigenda/api/clients/mock"
 	"github.com/Layr-Labs/eigenda/disperser"
 	"github.com/Layr-Labs/eigenda/tools/traffic"
 	"github.com/Layr-Labs/eigensdk-go/logging"
